make busybox config file not mandatory.

+++ scripts/bs_system_shell.sh
@@ -33,7 +33,15 @@ then
 
 		make ${NBCORE} ARCH=${KERNEL_ARCH} CROSS_COMPILE=${TGT_MACH}-  defconfig || exit 1
 		#make ${NBCORE}  ARCH=${KERNEL_ARCH}  CROSS_COMPILE=${TGT_MACH}-  menuconfig || exit 1
-		cp ${TARGET_CONFIG}/busybox_config .config || exit 1
+
+		if [ -f ${TARGET_CONFIG}/busybox_config ]
+		then
+		(
+			echo Busybox target config available...
+			# Use the target config
+			cp ${TARGET_CONFIG}/busybox_config .config || exit 1
+		)
+		fi
 
 		make ${NBCORE} ARCH=${KERNEL_ARCH} CROSS_COMPILE=${TGT_MACH}-  || exit 1
 		make ${NBCORE} ARCH=${KERNEL_ARCH} CROSS_COMPILE=${TGT_MACH}-  install CONFIG_PREFIX=${TARGET_ROOTFS} || exit 1
